fixes #168 UDP transport
This is the initial implementation of UDP transport.
It does in order guarantees (and consequently filters duplicates),
but it does not guarantee delivery.  The protocol limits payloads
to 65000 bytes (minus headers for SP), but you really want to
keep it to much less -- probably best for short messages that within
a single MTU to avoid IP fragmentation and reassembly.

This is unicast only for now (although there are plans for some
support for multicast and broadcast as well as being able to
perform automatic mesh building, but that will be in following work.

Additional tunables are coming.  This is only lightly tested at
this point, and should be considered experimental.  Its also undocumented.

+++ src/core/platform.h
@@ -356,7 +356,7 @@ extern int nni_tcp_listener_get(
 // Symbolic service names will be looked up assuming SOCK_STREAM, so
 // they may not work with UDP.
 extern void nni_resolv_ip(
-    const char *, const char *, int, bool, nng_sockaddr *sa, nni_aio *);
+    const char *, const char *, uint16_t, bool, nng_sockaddr *sa, nni_aio *);
 
 // nni_parse_ip parses an IP address, without a port.
 extern int nni_parse_ip(const char *, nng_sockaddr *);
